implements email notification. connects profile page to backend.
