+++ backend/.env
@@ -1,1 +1,7 @@
 JWT_SECRET=ourSecretisKey
+EMAIL_USER=J00952034@students.jsums.edu
+EMAIL_PASS=arsqdmcmsqmmauzw 
+DATABASE_URL=sqlite://./sensai.db
+API_SECRET=your_api_secret_key
+PORT=5001
+
