+++ backend/package.json
@@ -16,6 +16,7 @@
     "express": "^4.21.2",
     "jsonwebtoken": "^9.0.2",
     "mysql2": "^3.14.0",
+    "nodemailer": "^6.10.1",
     "sqlite3": "^5.1.7"
   }
 }
